requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Flask (for admin dashboard)
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL
aiosqlite==0.19.0  # Async SQLite for development

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utils
python-dotenv==1.0.0
httpx==0.26.0

# Date handling
python-dateutil==2.8.2
